+++ couponmanager/src/main/kotlin/dev/klevente/coupunch/couponmanager/coupon/Extensions.kt
@@ -1,6 +1,7 @@
 package dev.klevente.coupunch.couponmanager.coupon
 
 import dev.klevente.coupunch.couponmanager.product.Product
+import java.math.BigDecimal
 
 fun Coupon.hasMoreEligibleProductsOrAnyProductGroups() = eligibleProducts.size > 1 || eligibleProductGroups.isNotEmpty()
 
@@ -10,9 +11,9 @@ fun Reward.hasMoreRewardProductsOrAnyProductGroups() = products.size > 1 || prod
 
 fun Reward.hasMoreRewardProductGroupsOrAnyProducts() = productGroups.size > 1 || products.isNotEmpty()
 
-fun Coupon.isRedeemable(progress: Double) = getRedeemLevel(progress) > -1
+fun Coupon.isRedeemable(progress: BigDecimal) = getRedeemLevel(progress) > -1
 
-fun Coupon.getRedeemLevel(progress: Double) = rewards.fold(-1) { acc, reward ->
+fun Coupon.getRedeemLevel(progress: BigDecimal) = rewards.fold(-1) { acc, reward ->
     if (progress >= reward.threshold) acc + 1
     else acc
 }
@@ -27,7 +28,7 @@ fun Coupon.getMergedEligibleProducts(): Map<Product, Int?> {
 }
 
 fun Coupon.getProgressFor(product: Product) = when (type) {
-    CouponType.POINT -> getMergedEligibleProducts()[product]!!.toDouble()
+    CouponType.POINT -> getMergedEligibleProducts()[product]!!.toBigDecimal()
     CouponType.PRICE -> product.price
 }
 
@@ -40,11 +41,11 @@ fun Reward.getMergedProducts(): Map<Product, Int> {
     return mergedProducts
 }
 
-fun Map.Entry<Coupon, Double>.isRedeemable() = key.isRedeemable(value)
+fun Map.Entry<Coupon, BigDecimal>.isRedeemable() = key.isRedeemable(value)
 
-fun Map.Entry<Coupon, Double>.getRedeemLevel(): Int = key.getRedeemLevel(value)
+fun Map.Entry<Coupon, BigDecimal>.getRedeemLevel(): Int = key.getRedeemLevel(value)
 
-fun Product.calculateDiscountedPrice(discountType: DiscountType, discount: Double) = when (discountType) {
-    DiscountType.FIXED -> (price - discount).coerceAtLeast(0.0)
-    DiscountType.PERCENTAGE -> (price * (1 - discount / 100.0)).coerceAtLeast(0.0)
+fun Product.calculateDiscountedPrice(discountType: DiscountType, discount: BigDecimal) = when (discountType) {
+    DiscountType.FIXED -> (price - discount).coerceAtLeast(0.0.toBigDecimal())
+    DiscountType.PERCENTAGE -> (price * ((1).toBigDecimal() - discount / 100.0.toBigDecimal())).coerceAtLeast(0.0.toBigDecimal())
 }
